4_6_update

+++ view.h
@@ -6,7 +6,7 @@
 /*   By: jkimmina <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/04/05 13:11:35 by jkimmina          #+#    #+#             */
-/*   Updated: 2018/04/05 18:33:56 by jkimmina         ###   ########.fr       */
+/*   Updated: 2018/04/06 16:40:39 by jkimmina         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,6 @@
 # include "fdf.h"
 
 t_map	*rotate_x(t_map *map, int dgr);
-t_map	*rotate_y(t_map *map, double dgr);
+t_map	*rotate_y(t_map *map, int dgr);
 
 #endif
